backoffice: restart refactor
* ref: cern-sis/issues-inspire/issues/1539, cern-sis/issues-inspire/issues/1348
