--- backoffice/backoffice/authors/tests/cassettes/TestAuthorWorkflowViewSet.test_restart_current_task_without_failed_dags.yaml
@@ -1,187 +1,0 @@
-interactions:
-- request:
-    body: '{"dag_run_id": "00000000-0000-0000-0000-000000000000", "conf": {"workflow_id":
-      "00000000-0000-0000-0000-000000000000", "data": {"test": "test"}}}'
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '145'
-      Content-Type:
-      - application/json
-    method: POST
-    uri: http://airflow-api-server:8080/api/v2/dags/author_create_initialization_dag/dagRuns
-  response:
-    body:
-      string: "{\n  \"conf\": {\n    \"data\": {\n      \"test\": \"test\"\n    },\n
-        \   \"workflow_id\": \"00000000-0000-0000-0000-000000000000\"\n  },\n  \"dag_id\":
-        \"author_create_initialization_dag\",\n  \"dag_run_id\": \"00000000-0000-0000-0000-000000000000\",\n
-        \ \"data_interval_end\": \"2025-01-08T13:25:04.192208+00:00\",\n  \"data_interval_start\":
-        \"2025-01-08T13:25:04.192208+00:00\",\n  \"end_date\": null,\n  \"execution_date\":
-        \"2025-01-08T13:25:04.192208+00:00\",\n  \"external_trigger\": true,\n  \"last_scheduling_decision\":
-        null,\n  \"logical_date\": \"2025-01-08T13:25:04.192208+00:00\",\n  \"note\":
-        null,\n  \"run_type\": \"manual\",\n  \"start_date\": null,\n  \"state\":
-        \"queued\"\n}\n"
-    headers:
-      Cache-Control:
-      - no-store
-      Connection:
-      - close
-      Content-Length:
-      - '621'
-      Content-Type:
-      - application/json
-      Date:
-      - Wed, 08 Jan 2025 13:25:04 GMT
-      Server:
-      - gunicorn
-      X-Robots-Tag:
-      - noindex, nofollow
-    status:
-      code: 200
-      message: OK
-- request:
-    body: null
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-    method: GET
-    uri: http://airflow-api-server:8080/api/v2/dags/author_create_initialization_dag/dagRuns/00000000-0000-0000-0000-000000000000
-  response:
-    body:
-      string: "{\n  \"conf\": {\n    \"data\": {\n      \"test\": \"test\"\n    },\n
-        \   \"workflow_id\": \"00000000-0000-0000-0000-000000000000\"\n  },\n  \"dag_id\":
-        \"author_create_initialization_dag\",\n  \"dag_run_id\": \"00000000-0000-0000-0000-000000000000\",\n
-        \ \"data_interval_end\": \"2025-01-08T13:25:04.192208+00:00\",\n  \"data_interval_start\":
-        \"2025-01-08T13:25:04.192208+00:00\",\n  \"end_date\": null,\n  \"execution_date\":
-        \"2025-01-08T13:25:04.192208+00:00\",\n  \"external_trigger\": true,\n  \"last_scheduling_decision\":
-        null,\n  \"logical_date\": \"2025-01-08T13:25:04.192208+00:00\",\n  \"note\":
-        null,\n  \"run_type\": \"manual\",\n  \"start_date\": null,\n  \"state\":
-        \"queued\"\n}\n"
-    headers:
-      Cache-Control:
-      - no-store
-      Connection:
-      - close
-      Content-Length:
-      - '621'
-      Content-Type:
-      - application/json
-      Date:
-      - Wed, 08 Jan 2025 13:25:04 GMT
-      Server:
-      - gunicorn
-      X-Robots-Tag:
-      - noindex, nofollow
-    status:
-      code: 200
-      message: OK
-- request:
-    body: null
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-    method: GET
-    uri: http://airflow-api-server:8080/api/v2/dags/author_create_approved_dag/dagRuns/00000000-0000-0000-0000-000000000000
-  response:
-    body:
-      string: "{\n  \"detail\": \"DAGRun with DAG ID: 'author_create_approved_dag'
-        and DagRun ID: '00000000-0000-0000-0000-000000000000' not found\",\n  \"status\":
-        404,\n  \"title\": \"DAGRun not found\",\n  \"type\": \"https://airflow.apache.org/docs/apache-airflow/2.9.3/stable-rest-api-ref.html#section/Errors/NotFound\"\n}\n"
-    headers:
-      Cache-Control:
-      - no-store
-      Connection:
-      - close
-      Content-Length:
-      - '294'
-      Content-Type:
-      - application/problem+json
-      Date:
-      - Wed, 08 Jan 2025 13:25:04 GMT
-      Server:
-      - gunicorn
-      X-Robots-Tag:
-      - noindex, nofollow
-    status:
-      code: 404
-      message: NOT FOUND
-- request:
-    body: null
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-    method: GET
-    uri: http://airflow-api-server:8080/api/v2/dags/author_create_rejected_dag/dagRuns/00000000-0000-0000-0000-000000000000
-  response:
-    body:
-      string: "{\n  \"detail\": \"DAGRun with DAG ID: 'author_create_rejected_dag'
-        and DagRun ID: '00000000-0000-0000-0000-000000000000' not found\",\n  \"status\":
-        404,\n  \"title\": \"DAGRun not found\",\n  \"type\": \"https://airflow.apache.org/docs/apache-airflow/2.9.3/stable-rest-api-ref.html#section/Errors/NotFound\"\n}\n"
-    headers:
-      Cache-Control:
-      - no-store
-      Connection:
-      - close
-      Content-Length:
-      - '294'
-      Content-Type:
-      - application/problem+json
-      Date:
-      - Wed, 08 Jan 2025 13:25:04 GMT
-      Server:
-      - gunicorn
-      X-Robots-Tag:
-      - noindex, nofollow
-    status:
-      code: 404
-      message: NOT FOUND
-- request:
-    body: null
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '0'
-    method: DELETE
-    uri: http://airflow-api-server:8080/api/v2/dags/author_create_initialization_dag/dagRuns/00000000-0000-0000-0000-000000000000
-  response:
-    body:
-      string: ''
-    headers:
-      Cache-Control:
-      - no-store
-      Connection:
-      - close
-      Content-Type:
-      - application/json
-      Date:
-      - Wed, 08 Jan 2025 13:25:04 GMT
-      Server:
-      - gunicorn
-      X-Robots-Tag:
-      - noindex, nofollow
-    status:
-      code: 204
-      message: NO CONTENT
-version: 1
